Organize composables by responsibility. Minor improvements
